[feature #162477137] refactored test suite
