+++ .coveralls.yml
@@ -1,1 +1,2 @@
+
 repo_token: ubo0Len8MBcarGHsgbm2h34sQbkyW6Xph;
